requirements.txt
fastapi
python-dotenv
requests
uvicorn
google-generativeai
cmake
httpx
pinecone
